Add bash-loader
I have some environments that MyCmd will be deployed to where a new
enough Bash is not available on the PATH and it is not convenient to put
one on the PATH.

`bash-loader` is a wrapper that can be used with `/usr/bin/env` in a
shebang line. As long as the user has the `MYCMD_BASH_PATH` environment
variable set with a new enough Bash, this will properly pass any script
off to the Bash.

`bash-loader` itself will just need to be on the PATH itself somewhere.