Implement the `mycmd migrate` command group
Automated refactoring of existing command groups and commands to fit
with the new style of implementation. Still a bit more to do, but was
sufficient to migrate my personal command groups and commands. Next step
is to try on my much larger set of command groups and commands for work.