Further optimize command group loading
Skipping doing validation of parameters passed into
`register_command_group` if it is being called from
`mycmd:pathname.walk_path`, as the paths are validated there.

This is in lieu of facilities we would have in other languages to ensure
types with the correct parameters are passed in.

A little ugly, but this is a speed critical area.