Testing command help text
Realized that the reloading of the command and command group structs
within subshells, for example, in calling the get methods such as:

    "$(mycmd:command.get_help_text command)"

Since array variables are not exported, changes to them cannot be
reflected back to the parent Bash process.

Therefore, to enable not having to do repeated loads of metadata in the
common use cases of printing this data. Made a similar change to both
the command group and command structs.